Don't reconfigure if we're doing a dirty build & there's already a Makefile
By not re-running configure we avoid trashing the Makefile cache which was
causing even non-clean builder configurations to rebuild all the LLVM/Clang
source anyway.

This brings my internal GDB 7.5 tester down to under 10 minutes (on an
zero-change incremental rebuild), down from 16 minutes. A little over a minute
of that is running the Clang+LLVM tests and rebuilding GDB. Those could
potentially be optimized as well (skipping the tests entirely, and applying the
same conditional configure to the GDB build).