Add return statement to linpack main routine.
The "main" routine in linpack-pc.c is missing a return statement.
This doesn't matter when it is compiled with a C99 compiler, but
when built with a C89 compiler, it may cause the benchmark to
return with a random exit code even after successful execution.

Fixed by adding a "return 0;" statment.