Commit an initial Windows build to www/builds
This doesn't have to be the long term home for this file, and I'm assuming that
nobody is using git to maintain patches to www.  There are also many large
binary artifacts in devmtg already.

I also don't expect to ever have more than a few of these live at a time,
unlike the LLVM releases page.