Give performance builders more understandable names
We now use CPU architecture + clang flags to describe the builders and
also sort the non-polly builder next to the other clang builders.

This should allow people to better understand what hardware is tested
and especially avoid confusion that all builders are run with polly
enabled and consequently not useful for other people. We have a plain
'clang -O3' builder which is run in-parallel on five machines and which
should be useful for the wider audience (Also we could use some newer
hardware)